无人机飞行记录修改、实时数据返回修改

+++ src/main/java/com/uav/ops/dto/req/FlyHistoryReqDTO.java
@@ -57,7 +57,7 @@ public class FlyHistoryReqDTO {
     private Integer mileage;
 
     @ApiModelProperty(value = "飞行时间 s")
-    private Integer flyTime;
+    private Long flyTime;
 
     @ApiModelProperty(value = "视频地址")
     private String flyVideo;
